+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 New Features:
 
 - No Double Sneak
-- 
+- No Resource Pack Compatibility Waring